+++ src/index.ts
@@ -1,3 +1,4 @@
+export * from "./Component";
 export { default as Component } from "./Component";
 export { default as UUID } from "./UUID";
 export * from "./Decorators";