+++ CHANGES.md
@@ -1,5 +1,16 @@
 # Armature Change Log
 
+## 0.1.3
+- Added `Component#$getTypeName()`
+- Updated `Component#getIdentifier` to use `getTypeName`
+- Added documentation to most methods
+- Added test suite (Mocha, Chai, and Karma)
+	- Test coverage is incomplete
+- Removed 'debug' build path
+
+## 0.1.1, 0.1.2
+- Fixed typos and build regressions
+
 ## 0.1.0
 - Initial release
 - Based on internal prototype